Added methods that check if Mixers and Sounds exist. Sounds can now be paused and unpaused
Added DoesMixerExist, DoesMixerGroupExist, DoesSoundExist, and DoesSourcesSoundExist methods. Added DeciblesToPercent and PercentToDecibles methods to help convert volume levels for use with UI. Added Getters and Setters for MixerMasterVolume and MixerGroupVolume for use with UI. Added Pause, Unpause, PauseAllButMixerGroup, UnpauseAllButMixerGroup, PauseAllFromMixerGroup, UnpauseAllFromMixerGroup, PauseAll, and UnpauseAll methods. Reformatted code for performance and readability. Added MixerInfo class that stores a mixer and the updateMode for that mixer so it can be easily set up from the Inspector. Added GameAudioVersion to keep track of the version of audio. Added AudioVersion string in AudioData class. If an old save doesn't match it will not be loaded.